create token object
