--- lib/core/src/source_code/line.rs
@@ -1,11 +1,0 @@
-use crate::positive_integer::PositiveInteger64;
-
-pub struct Line {
-    pub number: PositiveInteger64,
-}
-
-impl Line {
-    pub fn new(number: PositiveInteger64) -> Self {
-        Self { number }
-    }
-}
