--- lib/core/src/token/token_type.rs
@@ -1,54 +1,0 @@
-#[derive(Debug, PartialEq, Clone)]
-pub enum TokenType {
-    DelimiterAngleLeft,
-    DelimiterAngleRight,
-    DelimiterBraceLeft,
-    DelimiterBraceRight,
-    DelimiterBracketLeft,
-    DelimiterBracketRight,
-    DelimiterParenthesisLeft,
-    DelimiterParenthesisRight,
-    DoubleEqual,
-    DoubleSlash,
-    Eof,
-    Identifiers(String),
-    KeywordAnd,
-    KeywordAs,
-    KeywordElse,
-    KeywordIf,
-    KeywordInterface,
-    KeywordMatch,
-    KeywordMutable,
-    KeywordNot,
-    KeywordOperationBreak,
-    KeywordOperationContinue,
-    KeywordOperationFor,
-    KeywordOperationLoop,
-    KeywordOperationReturn,
-    KeywordOr,
-    KeywordTypeBoolean,
-    KeywordTypeFloat,
-    KeywordTypeFunction,
-    KeywordTypeHash,
-    KeywordTypeInteger,
-    KeywordTypeList,
-    KeywordTypeNone,
-    KeywordTypeOption,
-    KeywordTypeStruct,
-    LiteralBoolean(bool),
-    LiteralCharacter(char),
-    LiteralFloat(f64),
-    LiteralInteger(i32),
-    LiteralNone,
-    LiteralString(String),
-    SeparatorColon,
-    SeparatorComma,
-    SeparatorDot,
-    SeparatorSemicolon,
-    SingleAsterisk,
-    SingleEqual,
-    SingleMinus,
-    SinglePercent,
-    SinglePlus,
-    SingleSlash,
-}
